feat: support r84

+++ src/main/java/io/luxus/adofai/converter/converters/effect/TransparentMapConverter.java
@@ -17,7 +17,7 @@ public class TransparentMapConverter implements MapConverter {
     @Override
     public Object[] prepareParameters(Scanner scanner) {
         System.out.print("투명도(0~100):");
-        int opacity = scanner.nextInt();
+        double opacity = scanner.nextDouble();
         scanner.nextLine();
 
         return new Object[] { opacity };
@@ -25,7 +25,7 @@ public class TransparentMapConverter implements MapConverter {
 
     @Override
     public boolean impossible(CustomLevel customLevel, Object... args) {
-        int opacity = (int) args[0];
+        double opacity = (double) args[0];
 
         if (opacity < 0 || opacity > 100) {
             System.err.println("투명도는 0이상 100이하여야합니다. (opacity=" + opacity + ")");
@@ -46,7 +46,7 @@ public class TransparentMapConverter implements MapConverter {
             return null;
         }
 
-        int opacity = (int) args[0];
+        double opacity = (double) args[0];
 
         return MapConverterBase.convert(customLevel, false,
                 applyEach -> applyEach.getOneTimingTiles().stream()
@@ -60,7 +60,7 @@ public class TransparentMapConverter implements MapConverter {
                 });
     }
 
-    private MoveTrack getTransparentMoveTrack(long opacity) {
+    private MoveTrack getTransparentMoveTrack(double opacity) {
         return new MoveTrack.Builder()
                 .startTilePosition(TilePosition.START).endTilePosition(TilePosition.END)
                 .duration(0.0)
@@ -68,7 +68,7 @@ public class TransparentMapConverter implements MapConverter {
                 .build();
     }
 
-    private void editOpacityEvents(Tile.Builder tileBuilder, long opacity) {
+    private void editOpacityEvents(Tile.Builder tileBuilder, double opacity) {
         tileBuilder.removeActions(EventType.ANIMATE_TRACK);
         tileBuilder.<MoveTrack>editActions(EventType.MOVE_TRACK, a -> new MoveTrack.Builder().from(a)
                 .opacity(a.getOpacity() == 0 ? 0 : opacity)
